+++ savvy-ffi/src/lib.rs
@@ -67,6 +67,7 @@ extern "C" {
 extern "C" {
     pub fn Rf_xlength(arg1: SEXP) -> R_xlen_t;
     pub fn Rf_allocVector(arg1: SEXPTYPE, arg2: R_xlen_t) -> SEXP;
+    pub fn Rf_install(arg1: *const ::std::os::raw::c_char) -> SEXP;
     pub fn Rf_getAttrib(arg1: SEXP, arg2: SEXP) -> SEXP;
     pub fn Rf_setAttrib(arg1: SEXP, arg2: SEXP, arg3: SEXP) -> SEXP;
 }