+++ savvy/src/sexp/mod.rs
@@ -1,4 +1,4 @@
-use std::ffi::CStr;
+use std::ffi::{CStr, CString};
 
 use savvy_ffi::{
     Rf_isInteger, Rf_isLogical, Rf_isReal, Rf_isString, Rf_type2char, EXTPTRSXP, SEXP, TYPEOF,
@@ -132,21 +132,109 @@ impl Sexp {
         }
     }
 
-    pub fn get_class(&self) -> Option<Vec<&'static str>> {
-        let class_sexp = unsafe { savvy_ffi::Rf_getAttrib(self.0, savvy_ffi::R_ClassSymbol) };
+    pub fn get_attrib(&self, attr: &str) -> crate::error::Result<Option<Sexp>> {
+        let attr_cstr = match CString::new(attr) {
+            Ok(cstr) => cstr,
+            Err(e) => return Err(crate::error::Error::new(&e.to_string())),
+        };
+        let attr_sexp = unsafe {
+            crate::unwind_protect(|| {
+                savvy_ffi::Rf_getAttrib(self.0, savvy_ffi::Rf_install(attr_cstr.as_ptr()))
+            })?
+        };
 
-        if class_sexp == unsafe { savvy_ffi::R_NilValue } {
+        if attr_sexp == unsafe { savvy_ffi::R_NilValue } {
+            Ok(None)
+        // Bravely assume the "class" attribute is always a valid STRSXP.
+        } else {
+            Ok(Some(Sexp(attr_sexp)))
+        }
+    }
+
+    unsafe fn get_string_attrib_by_symbol(&self, attr: SEXP) -> Option<Vec<&'static str>> {
+        let sexp = savvy_ffi::Rf_getAttrib(self.0, attr);
+
+        if sexp == unsafe { savvy_ffi::R_NilValue } {
             None
         // Bravely assume the "class" attribute is always a valid STRSXP.
         } else {
-            Some(crate::StringSexp(class_sexp).iter().collect())
+            Some(crate::StringSexp(sexp).iter().collect())
         }
     }
+
+    pub fn get_class(&self) -> Option<Vec<&'static str>> {
+        unsafe { self.get_string_attrib_by_symbol(savvy_ffi::R_ClassSymbol) }
+    }
+
+    pub fn get_names(&self) -> Option<Vec<&'static str>> {
+        unsafe { self.get_string_attrib_by_symbol(savvy_ffi::R_NamesSymbol) }
+    }
+
+    pub fn set_attrib(&mut self, attr: &str, value: Sexp) -> crate::error::Result<()> {
+        let attr_cstr = match CString::new(attr) {
+            Ok(cstr) => cstr,
+            Err(e) => return Err(crate::error::Error::new(&e.to_string())),
+        };
+        unsafe {
+            crate::unwind_protect(|| {
+                savvy_ffi::Rf_setAttrib(self.0, savvy_ffi::Rf_install(attr_cstr.as_ptr()), value.0)
+            })?
+        };
+
+        Ok(())
+    }
+
+    unsafe fn set_string_attrib_by_symbol(
+        &mut self,
+        attr: SEXP,
+        values: &[&str],
+    ) -> crate::error::Result<()> {
+        let values_sexp: Sexp = values.try_into()?;
+        unsafe { crate::unwind_protect(|| savvy_ffi::Rf_setAttrib(self.0, attr, values_sexp.0))? };
+
+        Ok(())
+    }
+
+    pub fn set_class(&mut self, classes: &[&str]) -> crate::error::Result<()> {
+        unsafe { self.set_string_attrib_by_symbol(savvy_ffi::R_ClassSymbol, classes) }
+    }
+
+    pub fn set_names(&mut self, names: &[&str]) -> crate::error::Result<()> {
+        unsafe { self.set_string_attrib_by_symbol(savvy_ffi::R_NamesSymbol, names) }
+    }
+}
+
+pub(crate) fn get_dim_from_sexp(value: SEXP) -> Option<Vec<usize>> {
+    let dim_sexp = unsafe { savvy_ffi::Rf_getAttrib(value, savvy_ffi::R_DimSymbol) };
+
+    if dim_sexp == unsafe { savvy_ffi::R_NilValue } {
+        None
+    // Bravely assume the "dim" attribute is always a valid INTSXP.
+    } else {
+        Some(
+            crate::IntegerSexp(dim_sexp)
+                .as_slice()
+                .iter()
+                .map(|i| *i as _)
+                .collect(),
+        )
+    }
+}
+
+pub(crate) fn set_dim_to_sexp(value: SEXP, dim: &[usize]) -> crate::error::Result<()> {
+    let dim_sexp: Sexp = dim
+        .iter()
+        .map(|v| *v as i32)
+        .collect::<Vec<i32>>()
+        .try_into()?;
+    unsafe { savvy_ffi::Rf_setAttrib(value, savvy_ffi::R_DimSymbol, dim_sexp.0) };
+    Ok(())
 }
 
 macro_rules! impl_common_sexp_ops {
     ($ty: ty) => {
         impl $ty {
+            #[inline]
             pub fn inner(&self) -> savvy_ffi::SEXP {
                 self.0
             }
@@ -155,42 +243,25 @@ macro_rules! impl_common_sexp_ops {
                 unsafe { savvy_ffi::Rf_xlength(self.inner()) as _ }
             }
 
+            #[inline]
             pub fn is_empty(&self) -> bool {
                 self.len() == 0
             }
 
-            pub fn get_names(&self) -> Vec<&'static str> {
-                let names_sexp =
-                    unsafe { savvy_ffi::Rf_getAttrib(self.inner(), savvy_ffi::R_NamesSymbol) };
+            pub fn get_attrib(&self, attr: &str) -> crate::error::Result<Option<Sexp>> {
+                crate::Sexp(self.inner()).get_attrib(attr)
+            }
 
-                if names_sexp == unsafe { savvy_ffi::R_NilValue } {
-                    std::iter::repeat("").take(self.len()).collect()
-                // Bravely assume the "name" attribute is always a valid STRSXP.
-                } else {
-                    crate::StringSexp(names_sexp).iter().collect()
-                }
+            pub fn get_names(&self) -> Option<Vec<&'static str>> {
+                crate::Sexp(self.inner()).get_names()
             }
 
             pub fn get_class(&self) -> Option<Vec<&'static str>> {
-                crate::Sexp(self.0).get_class()
+                crate::Sexp(self.inner()).get_class()
             }
 
             pub fn get_dim(&self) -> Option<Vec<usize>> {
-                let dim_sexp =
-                    unsafe { savvy_ffi::Rf_getAttrib(self.inner(), savvy_ffi::R_DimSymbol) };
-
-                if dim_sexp == unsafe { savvy_ffi::R_NilValue } {
-                    None
-                // Bravely assume the "dim" attribute is always a valid INTSXP.
-                } else {
-                    Some(
-                        crate::IntegerSexp(dim_sexp)
-                            .as_slice()
-                            .iter()
-                            .map(|i| *i as _)
-                            .collect(),
-                    )
-                }
+                crate::sexp::get_dim_from_sexp(self.inner())
             }
         }
     };
@@ -199,16 +270,51 @@ macro_rules! impl_common_sexp_ops {
 macro_rules! impl_common_sexp_ops_owned {
     ($ty: ty) => {
         impl $ty {
+            #[inline]
             pub fn inner(&self) -> SEXP {
                 self.inner
             }
 
+            #[inline]
             pub fn len(&self) -> usize {
                 self.len
             }
 
+            #[inline]
             pub fn is_empty(&self) -> bool {
                 self.len == 0
+            }
+
+            pub fn get_attrib(&self, attr: &str) -> crate::error::Result<Option<Sexp>> {
+                crate::Sexp(self.inner()).get_attrib(attr)
+            }
+
+            pub fn get_names(&self) -> Option<Vec<&'static str>> {
+                crate::Sexp(self.inner()).get_names()
+            }
+
+            pub fn get_class(&self) -> Option<Vec<&'static str>> {
+                crate::Sexp(self.inner()).get_class()
+            }
+
+            pub fn get_dim(&self) -> Option<Vec<usize>> {
+                crate::sexp::get_dim_from_sexp(self.inner())
+            }
+
+            pub fn set_attrib(&mut self, attr: &str, value: Sexp) -> crate::error::Result<()> {
+                crate::Sexp(self.inner()).set_attrib(attr, value)
+            }
+
+            pub fn set_class(&mut self, classes: &[&str]) -> crate::error::Result<()> {
+                crate::Sexp(self.inner()).set_class(classes)
+            }
+
+            pub fn set_names(&mut self, names: &[&str]) -> crate::error::Result<()> {
+                crate::Sexp(self.inner()).set_names(names)
+            }
+
+            pub fn set_dim(&mut self, dim: &[usize]) -> crate::error::Result<()> {
+                crate::sexp::set_dim_to_sexp(self.inner(), dim)
             }
         }
     };